sml: rework read() to be simpler
We need to read data only every 5-6min so instead of keep
the serial port open and risking data loss (or having to keep
reading data we don't need) this commit just switch to open
the port, read until we have a valid frame and close it again.

This is fine because the energy meter will push out data every
few second.